Do not regenerate tree renders on every query
The model should enforce that tree renders do not change for different
areas of the trace. The only reason it should change is if new data is
available (for example if the state system continues being built between
queries).

The view should still support the case where the tree render changes
though, so it should not be aware of what's happening at the model
level. It can however avoid repainting the tree items if they did not
change.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>