Strip new syscall prefixes in the state labels
Earlier versions of LTTng used "sys_" as prefix for syscall events,
which we stripped in the labels. Recent versions now use
"syscall_entry_", which we should also strip from the labels.

Note this value is present in the kernel event layout, ideally it should
be reused from there. We'd just need a way to match the CFV render
provider with its corresponding event layout (this must be specific to
the CFV render though!)

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>