Implement clearing the view
To clear the view programmatically, one would call
control.initializeForTrace(null).

Unfortunately the current signal manager sends both traceSelected and
traceClosed, in no particular order, when a trace is closed and a it
switches to a new one. This makes it very hard to differentiate between
"there is still a trace left, show that one" and "there are no trace
left, clear the view". Jabberwocky ought to fix this.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>