Convert the Eclipse toolbar to a JavaFX Toolbar
Use a JavaFX toolbar within a BorderPane to implement the view's
toolbar. This puts all the related components (Buttons, menu items,
event handlers etc.) under the viewer, ready to be moved to their own
widgets, which may not be a ViewPart directly.

MenuButton allows showing a label and a drop-down (and even an icon if
we wanted). There didn't seem to be any equivalent in SWT.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>