Fix repaintCurrentArea() method
Since the time range did not move, the painting task never asked for a
refresh, leaving the view empty. Add a flag to force a repaint, which
will then be requested when at every call to seekVisibleRange(), even if
we are "seeking" to the exact same time range we are already at.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>