timegraph: Remove padding between back/forward nav buttons
Wrap the buttons into an HBox, which allows them to use a different
padding value than the other buttons in the toolbar.

Also split the SplitMenuButton into a separate Button and MenuButton, it
was weird to have the Forward button have the same focus as the
drop-down, but not the Back one.
It also allows having a consistent separation between the 3 buttons,
without having to go crazy with CSS.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>