timegraph: Add utilities for testing state rendering
Add prepare/wait methods making use of a latch, similar to the one in
the control for waiting on signals. However, since the latch-countdown
is done in the UI thread, use await(timeout) and return a boolean
indicating if the latch was released. If not, we can run UI operations
again until it is counted down.

The usage pattern in tests would look like:

    getViewer().prepareWaitForRepaint();
    getViewer().paintCurrentLocation();
    while (!getViewer().waitForRepaint()) {
        updateUI();
    }

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>