timegraph: Add tooltips to state intervals
Drop the Supplier thing after all, while this could have been used to
defer processing to a later time, it then had the side-effect of keeping
many more data structures in memory, which made memory usage explode...

Fetch all the necessary data to populate the tooltip in advance, but
only create and install the tooltip UI objects on mouse enter. This
balance gives the best user experience in extreme scenarios.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>