timegraph: Add an interface for model state intervals
Using a Supplier for properties is just an implementation detail (that
our reference implementation happens to use). Split out an interface for
the return methods, allowing different implementations to do what they
want.

Start not using I- for interfaces, while at it. We should go over all of
them eventually.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>