timegraph: Implement hiding/showing individual arrow series
Adds a new MenuButton to the toolbar to list the available arrow series.
Selecting/deselecting them wil show/hide the corresponding arrows.

It's not just a matter of showing/hiding UI objects though: if a series
is hidden we can avoid its related queries altogether. On re-show
however, we need to fetch the data we didn't fetch before.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>