timegraph: Correctly reposition the text labels an arrows on zoom
There is still a small "flicker" for text labels that are clamped to the
side of the pane for states that begin earlier. This is because the
corrected position is not the clamped one anymore. It's not a big deal,
and if anything it helps visualize the zoom animation.

Fixes #26.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>