timegraph: Add the concept of state definitions
The model will provide state definitions it makes use of, as well as
default value for its UI representation (color, line thickness for now,
but more criteria could be added).

This will allow re-using the same definitions accross different
views/models. For example, the "Running in usermode" state can be
presented in both the Threads and the CPU view. If the user modifies
the UI representation of this state, it should apply to all views.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>