timegraph: Add method to match a trace event to a tree element
If drawn event renders are to be provided independently of model
providers, they still need a way to match back to the specific tree
elements of any particular timegraph.

The idea of a drawn event is to represent, most of the time, a trace
event (whereas most other timegraph constructs represent states).
The dependency chain is then:

  Trace Event -> TimeGraphEvent -> TimeGraphDrawnEvent

A trace event is generic, but a TimeGraphEvent is already bound to a
particular time graph. To keep the chain possible, the model provider
can offer a method to match any trace event to the corresponding
element of its tree (if there is one).

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>