timegraph: Have each layer own their parent group instead of "being" it
Go back to a composition model for timegraph layers. The important part
here is that the widget manages the groups, and passes them to the
layers (or rather, "layer providers"). This will allow us to move to a
pattern where the widget can manage several Panes/Groups, like tiles,
and pass them to the layers in succession so they can paint accordingly.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>