lttng: Use a consistent state value for hardware IRQ activity
Just use the same state value as the "CPU running IRQ" state. That way
they will use the same state definition (same state in the color
configuration dialog).

No need to store the CPU as the state value, the IRQ is modeled under a
CPU sub-tree already, so we could query which one using the parent
attribute.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>