timegraph: Disable the collapsing of all tree items
The setClip() work-around works for collapsing the tree root, but as
soon as any other item (other than the very last one) is collapsed, the
data becomes misaligned and shows very wrong information.

No information is much better than wrong information, so we will disable
the collapsing for now. Supporting it correctly will be tracked in #29.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>