timegraph: Fix view initialization problem
A race condition can happen between the main thread and a widget's
thread's first redraw, where the timegraph width is reported as zero.
This leaves the widget in a weird state. Instead clamp to a value much
higher than 1 so that we obtain a resolution that at least makes sense.

This will also allow the widget controls to be initialized independently
of a Timeline view.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>