Rename project.core plugin to "o.l.s.jabberwocky"
This will be a temporary plugin that will contain the
JabberwockyProjectManager for now. TMF needs to depend on this since for
now the TMF trace objects create the Jabberwocky projects. Once we have
a "Project View" that creates JW projects, we could flip that dependency
around and merge this manager into the Scope main plugin.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>