Remove Symbol Providers
There is no way for users to configure symbol providers anymore (the
only button was within the Callstack View), so there is no need to keep
this code.

This configuration should instead be project-wide. The eventual
replacement dialog would fit well in a "Project Settings..." menu option
of the eventual new project view.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>