Introduce a "navigation area" and its widgets
We'll add the concept of a navigation area widget, which will be
different from a regular time graph widget, in that it should be of
rather small size, shoud *not* be resizable, and should show the full
range of the trace project.

The full-range XY-charts would be a good example of navigation widgets,
rather than standard resizable timegraph widgets.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>