Add mouse scroll handlers to XY Charts
This will allow zooming in and out using ctrl+scroll, just like in
timegraph widgets.

Unfortunately, since the timegraph and XYChart widgets do not share the
same control objects, it does not seem trivial to re-use the same logic
between the two.
If it wasn't of the variable zoom pivot (zoom operation being centered
on the mouse position), we could implement this generically at the
application level, independently of the widget.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>