Do not enforce time control range limits when changed programmatically
The "limits" property is used to ensure a minimum time range is
respected so that the timeline views can continue to make sense (and
that the timeline and time control fields stay consistent).

However in very short traces, it is possible that the total project
length is shorter than this limit.

In these cases, we will let the framework set a lower limit to the time
control widgets. This will make all values get clamped to the project
length in this case. Timeline views wouldn't make sense for these
projects anyway.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>