Link the notion of project range to the timestamp formatting
The "hh:mm:ss.n" format will require a reference to a given project, so
that we can determine which date is implied from the project's full
trace range.

Note that this still won't work for traces/projects that span multiple
days, since there could be confusion. In these cases the shortened
timestamp format won't be available.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>