Use a final view context in Time Control
Just like other widgets, it is much more straightforward to have the
time control belong to a given view context for its whole lifetime. If
we want to use several contexts in parallel, we can just instantiate
another TimeControl instance (it could still be displayed as the "same"
view in the UI).

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>