Correctly update views according to timestamp format changes
The following displays of timestamps should be refreshed automatically
whenever the user selects a new timestamp format:
- Time Control project ranges
- Event table timestamps

Note the time control will show the dates for the project ranges even
while in hh:mm:ss format. This will be the only place in the application
that shows the date, so it will give some reference without having to
clutter every single timetamp display.

Fixes #35.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>