Hook the timegraph's redraw job to the context's project-switch freeze
The redraw task was currently fetching and using time ranges and project
references while project switching operations were ongoing. This is bad!
If it reads those values, then it should check the 'listenerFreeze' flag
first, and do nothing if it is on.

Another option could have been to stop scheduling of this thread in the
flush() method, and resume it in the project-change callback. But since
there is only 1 of those threads, it's not too costly to simply keep the
ping going and do a no-op.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>