Use a custom TimeAxis for visible-range XYCharts
This will allow computing the ticks ourselves, so that they align on
"round" numbers. We also won't be bound on having ticks at the start and
end all the time.

Timestamp labels are shown, but are not formatted at the moment.

Unfortunately we can't extend NumberAxis, so we have to copy it over.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>