Merge all LTTng Scope repositories into this one
Use Maven modules instead of having different repositories with
different Maven projects. Other dependencies are not used by other
projects anyway, so they don't need to be in their own repositories.

Having a single repository makes it easier to release/deploy and to
checkout a different version (and to bisect with Git).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>